Modified Windows impl of window.go in order to implement better frameless support.
Removes use of `dwmExtendFrameIntoClientArea` which causes rounded corners and shadowing to disappear. Frame extension is done via `WM_NCCALCSIZE` instead which works better than `dwmExtendFrameIntoClientArea`.

This patch adds rounded corners by default on Windows 11 when running the app in frameless mode.

This patch adds proper shadowing around the window on Windows 11 and Windows 10.

Issue(s):
Double click maximize is not functional. I have not experimented why, but this can be solved in javascript rather than in the window code.